# backend/docs/swagger.yaml
basePath: /api
definitions:
  ds.CartBadgeDTO:
    properties:
      count:
        type: integer
      diet_id:
        type: integer
    type: object
  ds.DietDTO:
    properties:
      PGP:
        type: number
      PRP:
        type: number
      c_pol:
        type: integer
      complition_date:
        type: string
      creation_date:
        type: string
      creator_login:
        type: integer
      forming_date:
        type: string
      id:
        type: integer
      moderator_login:
        type: integer
      n_pol:
        type: integer
      products:
        items:
          $ref: '#/definitions/ds.ProductInDietDTO'
        type: array
      status:
        type: integer
    type: object
  ds.DietResolveRequest:
    properties:
      action:
        description: '"complete" | "reject"'
        type: string
    required:
    - action
    type: object
  ds.DietUpdateRequest:
    properties:
      PGP:
        type: number
      PRP:
        type: number
      c_pol:
        type: integer
      n_pol:
        type: integer
    required:
    - c_pol
    - n_pol
    type: object
...
  ds.ProductInDietDTO:
    properties:
      c_pol:
        type: integer
      description:
        type: string
      image:
        type: string
      n_pol:
        type: integer
      product_id:
        type: integer
      text:
        type: string
      title:
        type: string
    type: object
  ds.ProductToDietUpdateRequest:
    properties:
      description:
        type: string
    type: object
...
host: localhost:8090
info:
  contact:
    email: support@example.com
    name: API Support
  description: API-сервер для управления заявками и продуктами в системе DIET.
  title: API для системы DIET
  version: "1.0"
paths:
...
  /diet:
    get:
      description: Возвращает отфильтрованный список всех сформированных заявок (кроме
        черновиков и удаленных).
      parameters:
      - description: Фильтр по статусу заявки
        in: query
        name: status
        type: integer
      - description: Фильтр по дате 'от' (формат YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: Фильтр по дате 'до' (формат YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.DietDTO'
            type: array
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить список заявок (авторизованный пользователь)
      tags:
      - diet
  /diet/{id}:
    delete:
      description: Логически удаляет заявку, переводя ее в статус "удалена".
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить заявку (авторизованный пользователь)
      tags:
      - diet
    get:
      description: Возвращает полную информацию о заявке, включая привязанные продукты.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.DietDTO'
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить одну заявку по ID (авторизованный пользователь)
      tags:
      - diet
    put:
      consumes:
      - application/json
      description: Позволяет пользователю обновить поля своей заявки.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/ds.DietUpdateRequest'
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить данные заявки (авторизованный пользователь)
      tags:
      - diet
  /diet/{id}/form:
    put:
      description: Переводит заявку из статуса "черновик" в "сформирована".
      parameters:
      - description: ID заявки (черновика)
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Не все поля заполнены
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Сформировать заявку (авторизованный пользователь)
      tags:
      - diet
  /diet/{id}/products/{product_id}:
    delete:
      description: Удаляет связь между заявкой и продуктом.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: ID продукта
        in: path
        name: product_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить продукт из заявки (авторизованный пользователь)
      tags:
      - m-m
    put:
      consumes:
      - application/json
      description: Изменяет дополнительное описание для конкретного продукта в рамках
        одной заявки.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: ID продукта
        in: path
        name: product_id
        required: true
        type: integer
      - description: Новое описание
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/ds.ProductToDietUpdateRequest'
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить описание продукта в заявке (авторизованный пользователь)
      tags:
      - m-m
  /diet/{id}/resolve:
    put:
      consumes:
      - application/json
      description: Модератор завершает (с расчетом) или отклоняет заявку.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: 'Действие: ''complete'' или ''reject'''
        in: body
        name: action
        required: true
        schema:
          $ref: '#/definitions/ds.DietResolveRequest'
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Завершить или отклонить заявку (только модератор)
      tags:
      - diet
...
  /diet/productscart:
    get:
      description: Возвращает ID черновика текущего пользователя и количество продуктов
        в нем.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.CartBadgeDTO'
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить информацию для иконки корзины (авторизованный пользователь)
      tags:
      - diet
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"